feature:提供键过期数据统计

+++ petrichor/starter/src/main/java/com/jalinyiel/petrichor/start/domain/HotSpotDataInfoSummary.java
@@ -11,10 +11,10 @@ import java.util.List;
 @Builder
 public class HotSpotDataInfoSummary implements Serializable {
 
-    List<Double> hotSpotMemories;
+    List<Long> hotSpotMemories;
 
     List<String> hotSpotKeys;
 
-    List<Integer> hotSpotQueryTimes;
+    List<Long> hotSpotQueryTimes;
 
 }
